Refactor temperature logic and update thresholds
Revised `GetHeatOffset` in `DeviceSettingsFactory.cs` to adjust heat offset calculation by removing the previous logic and introducing a new condition that decreases the result by 2 if the exhaust air temperature is 22 or higher and the price level is not `EnergyPriceLevel.Extreme`.

Updated `GetComfortMode` to use `IsNightTime()` or `IsMidDay()` for determining `ComfortMode.Economy`.

Renamed `AdjustTempratureForTimeOfDay` to `AdjustFloorTempratureForTimeOfDay` in `DeviceSettingsFactory.cs`.

Modified `AdjustVentilationForEvening` in `HouseAutomationService.cs` to increase the exhaust air temperature threshold from 24 to 25 and the temperature difference threshold from 5 to 6.